feat: implement log line highlighting functionality for improved readability
