+++ assets/js/pages/bosun/index.vue
@@ -7,6 +7,7 @@ import Tooltip from '@/components/Tooltip.vue'
 import { highlightSQL } from '@/lib/highlightSQL'
 import { highlightJSON } from '@/lib/highlightJSON'
 import { highlightJS } from '@/lib/highlightJS'
+import { highlightLogLine } from '@/lib/highlightLog'
 import SlippyLoader from '@/components/SlippyLoader.vue'
 
 defineOptions({
@@ -386,43 +387,6 @@ let logsEventSource = null
 const logContainer = ref(null)
 const autoScroll = ref(true)
 
-function highlightLogLine(line) {
-  let s = line.replace(/&/g, '&amp;').replace(/</g, '&lt;').replace(/>/g, '&gt;')
-
-  // ISO timestamps
-  s = s.replace(/(\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}\.\d+Z?)/, '<span class="text-zinc-500">$1</span>')
-
-  // Log levels
-  s = s.replace(/\b(error|Error|ERROR|ERR)\b/g, '<span class="text-rose-500 font-semibold">$1</span>')
-  s = s.replace(/\b(warn|Warn|WARN|warning|Warning|WARNING)\b/g, '<span class="text-amber-500 font-semibold">$1</span>')
-  s = s.replace(/\b(info|Info|INFO)\b/g, '<span class="text-sky-500">$1</span>')
-  s = s.replace(/\b(debug|Debug|DEBUG|verbose|silly)\b/g, '<span class="text-zinc-500">$1</span>')
-
-  // HTTP methods
-  s = s.replace(/\b(GET|POST|PUT|PATCH|DELETE|HEAD|OPTIONS)\b/g, '<span class="text-cyan-500 font-medium">$1</span>')
-
-  // HTTP status codes (single-pass to avoid matching numbers in generated class names)
-  s = s.replace(/\b([2-5]\d{2})\b/g, (match, code) => {
-    const n = parseInt(code)
-    if (n >= 200 && n < 300) return `<span class="text-emerald-500">${code}</span>`
-    if (n >= 300 && n < 400) return `<span class="text-sky-500">${code}</span>`
-    if (n >= 400 && n < 500) return `<span class="text-amber-500">${code}</span>`
-    if (n >= 500 && n < 600) return `<span class="text-rose-500">${code}</span>`
-    return match
-  })
-
-  // API paths and URLs
-  s = s.replace(/(\/api\/[^\s"'<>]+)/g, '<span class="text-violet-500">$1</span>')
-  s = s.replace(/(https?:\/\/[^\s"'<>]+)/g, '<span class="text-sky-500 underline">$1</span>')
-
-  // Stack trace markers
-  s = s.replace(/(\s+at\s+)/g, '<span class="text-zinc-600">$1</span>')
-
-  // Brackets/tags
-  s = s.replace(/\[([^\]]+)\]/g, '<span class="text-zinc-600">[$1]</span>')
-
-  return s
-}
 
 function connectLogs() {
   if (logsEventSource) return
